requirements.txt
markdown
jinja2
pyyaml